requirements.txt
streamlit==1.31.0
transformers==4.37.2
spotipy==2.23.0
python-dotenv==1.0.1
torch==2.2.0